# fill paths faster, random stuff