before
`svg decode ........................ 19.724 ms     19.967 ms    ±0.195   x250`
after:
`svg decode ........................ 18.872 ms     19.060 ms    ±0.194   x262`